Add tags and titles as CLI opts
The major pain point about adding titles and notes through the editor is
that I need to build completion for it. I'm comfortable with completion
in Fish, not in Neovim. Therefore I copied code from fish-journal to add
titles and tags. It's kind of important being able to use consistent
tags. If I ever want to link notes through titles that also needs to be
consistent